requirements.txt
opencv-python
imagehash
Pillow
numpy
librosa
requests
beautifulsoup4
python-dotenv
google-generativeai
